제품 관련 페이지 추가

+++ PodoDemo/Controllers/ProductsController.cs
@@ -6,67 +6,166 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using PodoDemo.Models;
+using Microsoft.AspNetCore.Http;
+using Newtonsoft.Json;
 
 namespace PodoDemo.Controllers
 {
     public class ProductsController : Controller
     {
         private readonly PodoDemoNContext _context;
+        private static UserAuth _userAuth = new UserAuth();
 
         public ProductsController(PodoDemoNContext context)
         {
             _context = context;    
         }
 
-        // GET: Products
-        public async Task<IActionResult> Index()
+        /// <summary>
+        /// 사용자 권한 넣기
+        /// </summary>
+        /// <returns></returns>
+        public IActionResult CreaetUserAuth()
         {
-            return View(await _context.Product.ToListAsync());
+            CommonAPIController ss = new CommonAPIController(_context);
+            string userid = HttpContext.Session.GetString("userId");
+
+            // 사용자 세션 체크
+            if (!string.IsNullOrEmpty(userid))
+            {
+                _userAuth = ss.CheckUseauth(userid, "1-1");
+                return null;
+            }
+            else
+            {
+                return RedirectToAction("Error", "Home", new { errormessage = "UserauthError" });
+            }
         }
 
-        // GET: Products/Details/5
-        public async Task<IActionResult> Details(long? id)
+        /// <summary>
+        /// 목록 페이지로 이동
+        /// </summary>
+        /// <param name="isPop"></param>
+        /// <returns></returns>
+        public async Task<IActionResult> Index(bool? isPop)
         {
-            if (id == null)
+            CreaetUserAuth();
+
+            // 사용자 읽기 권한 체크
+            if (_userAuth.Read.Equals("4-3"))
             {
-                return NotFound();
+                return RedirectToAction("Error", "Home", new { errormessage = "UserauthError" });
             }
 
-            var product = await _context.Product
-                .SingleOrDefaultAsync(m => m.Productid == id);
-            if (product == null)
+            // 권한
+            ViewData["Read"] = _userAuth.Read;
+            ViewData["Write"] = _userAuth.Write;
+            ViewData["Modify"] = _userAuth.Modify;
+            ViewData["Delete"] = _userAuth.Delete;
+
+            if (isPop == null)
             {
-                return NotFound();
+                ViewBag.isPop = false;
+            }
+            else
+            {
+                ViewBag.isPop = isPop;
             }
 
-            return View(product);
+            List<Product> productList = await _context.Product.ToListAsync();
+
+            // 옵션 관련 이름으로 변환
+            foreach (Product item in productList)
+            {
+                if(item.Maker != null)
+                {
+                    item.Maker = _context.OptionMasterDetail.Where(x => x.Optionid == item.Maker).Single().Name;
+                }
+                item.Ownerid = _context.User.Single(x => x.Id == item.Ownerid).Name;
+            }
+
+            return View((Object)JsonConvert.SerializeObject(productList));
         }
 
-        // GET: Products/Create
+        
+        /// <summary>
+        /// 제품 생성 페이지로 이동
+        /// </summary>
+        /// <returns></returns>
         public IActionResult Create()
         {
+            // 사용자 권한
+            CreaetUserAuth();
+            ViewData["Read"] = _userAuth.Read;
+            ViewData["Write"] = _userAuth.Write;
+            ViewData["Modify"] = _userAuth.Modify;
+            ViewData["Delete"] = _userAuth.Delete;
+
+            // 사용자에게 쓰기 권한이 있는지 체크
+            if (_userAuth.Write.Equals("4-3"))
+            {
+                return RedirectToAction("Error", "Home", new { errormessage = "UserauthError" });
+            }
+
             return View();
         }
 
-        // POST: Products/Create
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
+        /// <summary>
+        /// 제품 실제 생성
+        /// </summary>
+        /// <param name="product"></param>
+        /// <returns></returns>
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Productid,Name,Maker,Ownerid,Origin,Productcode,Createdate,Createuser,Modifydate,Modifyuser")] Product product)
+        public async Task<IActionResult> Create([Bind("Name,Maker,Origin,Productcode")] Product product)
         {
+            // 사용자 쓰기 권한 체크
+            CreaetUserAuth();
+            if (_userAuth.Write.Equals("4-3"))
+            {
+                return RedirectToAction("Error", "Home", new { errormessage = "UserauthError" });
+            }
+
             if (ModelState.IsValid)
             {
+                product.Createdate = DateTime.Now;
+                product.Createuser = HttpContext.Session.GetString("userId");
+                product.Modifydate = DateTime.Now;
+                product.Modifyuser = HttpContext.Session.GetString("userId");
+                product.Ownerid = HttpContext.Session.GetString("userId");
+
                 _context.Add(product);
                 await _context.SaveChangesAsync();
                 return RedirectToAction("Index");
             }
+
+            #region 생성에 실패할 경우
+            ViewData["Read"] = _userAuth.Read;
+            ViewData["Write"] = _userAuth.Write;
+            ViewData["Modify"] = _userAuth.Modify;
+            ViewData["Delete"] = _userAuth.Delete;
+            #endregion
             return View(product);
         }
 
-        // GET: Products/Edit/5
+        /// <summary>
+        /// 수정 페이지로 이동
+        /// </summary>
+        /// <param name="id"></param>
+        /// <returns></returns>
         public async Task<IActionResult> Edit(long? id)
         {
+            // 읽기 권한이 없으면 아예 들어가지 못하게 한다.
+            CreaetUserAuth();
+            if (_userAuth.Read.Equals("4-3"))
+            {
+                return RedirectToAction("Error", "Home", new { errormessage = "UserauthError" });
+            }
+            ViewData["Read"] = _userAuth.Read;
+            ViewData["Write"] = _userAuth.Write;
+            ViewData["Modify"] = _userAuth.Modify;
+            ViewData["Delete"] = _userAuth.Delete;
+
             if (id == null)
             {
                 return NotFound();
@@ -80,13 +179,23 @@ namespace PodoDemo.Controllers
             return View(product);
         }
 
-        // POST: Products/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
+        /// <summary>
+        /// 제품 수정 기능 수행
+        /// </summary>
+        /// <param name="id"></param>
+        /// <param name="product"></param>
+        /// <returns></returns>
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(long id, [Bind("Productid,Name,Maker,Ownerid,Origin,Productcode,Createdate,Createuser,Modifydate,Modifyuser")] Product product)
         {
+            // 수정 권한 검사
+            CreaetUserAuth();
+            if (_userAuth.Modify.Equals("4-3"))
+            {
+                return RedirectToAction("Error", "Home", new { errormessage = "UserauthError" });
+            }
+
             if (id != product.Productid)
             {
                 return NotFound();
@@ -96,6 +205,9 @@ namespace PodoDemo.Controllers
             {
                 try
                 {
+                    product.Modifydate = DateTime.Now;
+                    product.Modifyuser = HttpContext.Session.GetString("userId");
+
                     _context.Update(product);
                     await _context.SaveChangesAsync();
                 }
@@ -112,32 +224,31 @@ namespace PodoDemo.Controllers
                 }
                 return RedirectToAction("Index");
             }
-            return View(product);
-        }
 
-        // GET: Products/Delete/5
-        public async Task<IActionResult> Delete(long? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var product = await _context.Product
-                .SingleOrDefaultAsync(m => m.Productid == id);
-            if (product == null)
-            {
-                return NotFound();
-            }
+            ViewData["Read"] = _userAuth.Read;
+            ViewData["Write"] = _userAuth.Write;
+            ViewData["Modify"] = _userAuth.Modify;
+            ViewData["Delete"] = _userAuth.Delete;
 
             return View(product);
         }
-
-        // POST: Products/Delete/5
+        
+        /// <summary>
+        /// 제품 삭제 
+        /// </summary>
+        /// <param name="id"></param>
+        /// <returns></returns>
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(long id)
+        public async Task<IActionResult> Delete(long id)
         {
+            // 권한 검사
+            CreaetUserAuth();
+            if (_userAuth.Delete.Equals("4-3"))
+            {
+                return RedirectToAction("Error", "Home", new { errormessage = "UserauthError" });
+            }
+
             var product = await _context.Product.SingleOrDefaultAsync(m => m.Productid == id);
             _context.Product.Remove(product);
             await _context.SaveChangesAsync();
